Cleanup StringHelper
