+++ selcukes-commons/src/main/java/io/github/selcukes/commons/http/WebResponse.java
@@ -19,7 +19,7 @@ package io.github.selcukes.commons.http;
 import com.fasterxml.jackson.databind.JsonNode;
 import io.github.selcukes.commons.exception.SelcukesException;
 import io.github.selcukes.databind.DataMapper;
-import io.github.selcukes.databind.utils.StringHelper;
+import io.github.selcukes.databind.utils.JsonUtils;
 import io.github.selcukes.databind.xml.XmlMapper;
 import lombok.Getter;
 import org.w3c.dom.Document;
@@ -191,7 +191,7 @@ public class WebResponse {
      * @return A JsonNode object
      */
     public JsonNode bodyJson() {
-        return StringHelper.toJson(httpResponse.body());
+        return JsonUtils.toJson(httpResponse.body());
     }
 
     /**
