+++ selcukes-core/src/main/java/io/github/selcukes/core/driver/AppiumOptions.java
@@ -19,8 +19,8 @@ package io.github.selcukes.core.driver;
 import io.appium.java_client.android.options.UiAutomator2Options;
 import io.appium.java_client.ios.options.XCUITestOptions;
 import io.appium.java_client.windows.options.WindowsOptions;
+import io.github.selcukes.collections.StringHelper;
 import io.github.selcukes.commons.helper.Preconditions;
-import io.github.selcukes.databind.utils.StringHelper;
 import lombok.experimental.UtilityClass;
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.MutableCapabilities;
