+++ selcukes-core/src/main/java/io/github/selcukes/core/driver/RunMode.java
@@ -20,7 +20,7 @@ import io.github.selcukes.commons.config.ConfigFactory;
 import io.github.selcukes.commons.os.Platform;
 import lombok.experimental.UtilityClass;
 
-import static io.github.selcukes.databind.utils.StringHelper.isNonEmpty;
+import static io.github.selcukes.collections.StringHelper.isNonEmpty;
 
 @UtilityClass
 public class RunMode {
