+++ selcukes-excel-runner/src/main/java/io/github/selcukes/excel/AbstractExcelDataProvider.java
@@ -17,9 +17,9 @@
 package io.github.selcukes.excel;
 
 import io.github.selcukes.collections.DataTable;
+import io.github.selcukes.collections.StringHelper;
 import io.github.selcukes.commons.exception.ExcelConfigException;
 import io.github.selcukes.databind.excel.ExcelMapper;
-import io.github.selcukes.databind.utils.StringHelper;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
