+++ selcukes-extent-reports/src/main/java/io/github/selcukes/extent/report/ExtentService.java
@@ -26,8 +26,8 @@ import com.aventstack.extentreports.reporter.configuration.ViewName;
 import io.github.selcukes.collections.Clocks;
 import io.github.selcukes.collections.Maps;
 import io.github.selcukes.collections.Resources;
+import io.github.selcukes.collections.StringHelper;
 import io.github.selcukes.databind.properties.PropertiesMapper;
-import io.github.selcukes.databind.utils.StringHelper;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -36,9 +36,9 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 import static io.github.selcukes.collections.Clocks.DATE_TIME_FILE_FORMAT;
+import static io.github.selcukes.collections.StringHelper.isNonEmpty;
 import static io.github.selcukes.commons.properties.SelcukesTestProperties.THUMBNAIL_REPORT;
 import static io.github.selcukes.commons.properties.SelcukesTestProperties.TIMESTAMP_REPORT;
-import static io.github.selcukes.databind.utils.StringHelper.isNonEmpty;
 import static java.util.Optional.ofNullable;
 
 public class ExtentService {
